requirements/base.txt
pydantic==2.6.4
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
python-multipart==0.0.9
alembic==1.13.2
sqlalchemy==2.0.31
sqlalchemy-filters==0.13.0
sqlalchemy-utils==0.14.2
python-slugify==8.0.4
slowapi==0.1.9
